requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu118

numpy>=1.19.0,<2.0
torchvision==0.15.2+cu118
torchaudio==2.0.2+cu118

numpy>=1.19.0
Pillow>=8.0.0
tensorboard>=2.4.0
tqdm
onnxruntime-gpu>=1.16.0